setUser should check if user data is already an object
It wasn't doing that if you passed it an object, it would only do it if
you gave it a function that returned an object. We're calling the
user function earlier in the stack so we can give it the request object
so it's already been evaluated at this point.