Add sending of custom data with `expressHandler`
Allows the sending of custom data along with requests handled by the `raygunClient.expressHandler`. Users can set the value `raygunClient.expressCustomData` to a function, and that function will be passed the current error, as well as the current request being processed. The user can then return an object that will be passed along as the custom data for the function.

Includes additional section in the README as well as a test for it.