Replace node-compass with node-sass-middleware
Node compass requires you to install a Ruby gem. There is no guarantee
that your average Node dev will have a functional Ruby toolchain or a
desire to install one.

The installation instructions do not mention the need for this Gem, so
the user would likely encounter a "spawn ENOENT compass" error, which is
a bit cryptic if you have no idea what compass is!

Additionally, this middleware is little-used in the node ecosystem, whereas
node-sass-middleware is well known and well used.

This commit also adds a package-lock.json, which is now standard for
Node apps.