Fix transport errors in tests
`onBeforeSend` expects a payload to be returned, and if it's not the
transport logs an error due to the message being undefined.

If we return the original payload the transport errors go away, but then
we get errors that `t.end` was called twice. We can instead just rely on
the `t.end` that gets called after the error is sent.