Update tap, add package-lock.json.
Turns out every single one of the 74 vulns in npm audit were due to our
old version of tap. Since it was our test suite, this isn't a huge
issue, but we might as well update to avoid audit notices.

This commit also adds a package-lock.json, which will help devs working
on the project ensure they're using the same dependency versions.