Tag unhandled exceptions from express middleware
Previously, this provider made no differentiation between handled
errors (where a user directly reports a caught error with .send) and an
unhandled error, where an indirect mechanism like the express error
handler is used to catch and report an error.

Since the express middleware is currently the only source of unhandled
errors, we can start by adding the "UnhandledException" tag to all
errors caught by the express middleware.

The bulk of this change adds a test to exercise the express middleware.
Other tests make use of `nock` to stop the tests from hitting the API,
but this is a tad inelegant for our case as we want to allow local
traffic, and ideally find out what the raygun API server would be
receiving.

The `makeClientWithMockServer` will create an instance of the Raygun
client, as well as an express server that mimics the crash entries
endpoints. This means we can use the client and any errors generated
will be submitted to our mock server, which we can then inspect
directly.

This is nice because it's the first test that actually verifies that the
data is submitted to the server in the expected format. Other tests
merely check that the message has been constructed correctly, but are
still vulnerable to transport failures.