Support batched error reporting
You can now pass {batch: true} when calling `.init` to enable the batch
transport.

When this option is enabled, errors will be placed in a queue and
processed at a regular interval.

This should cut down on transmission overhead if there are a large
number of errors in short succession.

You can further control the frequency at which batches are processed by
providing a `batchFrequency` option when initializing, which specifies
the interval between batches in milliseconds.