Ensure transport is present in OfflineStorage
Previously, we took the transport as an argument to `.init`. This meant
that it was possible for the transport to be undefined if we attempted
to send a message from the offline transport before calling `.init`.

Now, the transport is passed through the constructor, so it will always
be present assuming that the callsite is using strict mode typescript,
which we are.