Fix raygunClient.user not being called if request is missing
This is a fix for a regression I introduced during the TypeScript
overhaul.

Previously we would always call this.user, even if the request was null
or undefined. The regressed behavior would only call this.user if the
request was present, either from the middleware or if it was passed
manually.

This commit reverts the behavior back to what it was previously, and
adds a unit test to guard against future unintended changes to this
behaviour.