Fix exception formatting when passing string to send
Previously, when calling raygunClient.send("test"), the exception would
be formatted as if we had sent raygunClient.send(["t", "e", "s", "t"]).

This was as a result of passing the error value to humanString, which is
intending for serializing objects and arrays in a pleasant way.

This change avoids applying any formatting to exceptions passed as a
string.