Fix batch transport keeping process alive
Previously, using batch: true would keep the process alive indefinitely
due to the use of setInterval.

This is often okay for webapps which run until killed or crashed, but
causes problems when using crash reporting in a short-lived script.

This commit replaces the interval based process loop with one based on
setInterval.

The timer is now inactive until messages are queued for sending, at
which point we activate the timer loop.

Process will send batches, and restart the loop if there are any
remaining messages. If not, the loop will stop until new messages come
in.